+++ src/components/Typography/OverflowText.tsx
@@ -1,8 +1,9 @@
 import React from "react";
 
 import { CLASSPREFIX as eccgui } from "../../configuration/constants";
+import { TestableComponent } from "../../components/interfaces";
 
-export interface OverflowTextProps {
+export interface OverflowTextProps extends React.HTMLAttributes<HTMLElement>, TestableComponent {
     /**
      * How is ellipsis used to cut text overflows.
      * Use `reverse`to use the ellipis on text start and display the end of the text.
@@ -24,6 +25,10 @@ export interface OverflowTextProps {
      * HTML element that is used for the component.
      */
     useHtmlElement?: "p" | "div" | "span";
+    /**
+     * Used for all other necessary properties.
+     * @deprecated (v25) we will allow only basic HTML element properties and testing IDs
+     */
     [key: string]: any;
 }